+++ docxtpl.go
@@ -13,6 +13,7 @@ import (
 	"github.com/tomwatkins1994/go-docx-template/internal/contenttypes"
 	"github.com/tomwatkins1994/go-docx-template/internal/functions"
 	"github.com/tomwatkins1994/go-docx-template/internal/tags"
+	"github.com/tomwatkins1994/go-docx-template/internal/templatedata"
 )
 
 type DocxTmpl struct {
@@ -220,4 +221,61 @@ func (d *DocxTmpl) getDocumentXml() (string, error) {
 	}
 
 	return string(out), err
+}
+
+func (d *DocxTmpl) processTemplateData(data any) (map[string]any, error) {
+	convertedData, err := templatedata.DataToMap(data)
+	if err != nil {
+		return nil, err
+	}
+
+	var processTagValues func(data *map[string]any) error
+	processTagValues = func(data *map[string]any) error {
+		for key, value := range *data {
+			if stringVal, ok := value.(string); ok {
+				// Check for files
+				if isImage, err := templatedata.IsImageFilePath(stringVal); err != nil {
+					return err
+				} else {
+					if isImage {
+						image, err := CreateInlineImage(stringVal)
+						if err != nil {
+							return err
+						}
+						imageXml, err := d.addInlineImage(image)
+						if err != nil {
+							return err
+						}
+						(*data)[key] = imageXml
+					}
+				}
+			} else if nestedMap, ok := value.(map[string]any); ok {
+				if err := processTagValues(&nestedMap); err != nil {
+					return err
+				}
+				(*data)[key] = nestedMap
+			} else if sliceValue, ok := value.([]map[string]any); ok {
+				for i := range sliceValue {
+					if err := processTagValues(&sliceValue[i]); err != nil {
+						return err
+					}
+				}
+			} else if inlineImage, ok := value.(*InlineImage); ok {
+				imageXml, err := d.addInlineImage(inlineImage)
+				if err != nil {
+					return err
+				}
+				(*data)[key] = imageXml
+			}
+		}
+
+		return nil
+	}
+
+	err = processTagValues(&convertedData)
+	if err != nil {
+		return nil, err
+	}
+
+	return convertedData, nil
 }
