+++ internal/templatedata/template_data_processing.go
@@ -1,4 +1,4 @@
-package docxtpl
+package templatedata
 
 import (
 	"os"
@@ -6,64 +6,7 @@ import (
 	"slices"
 )
 
-func (d *DocxTmpl) processTemplateData(data any) (map[string]any, error) {
-	convertedData, err := dataToMap(data)
-	if err != nil {
-		return nil, err
-	}
-
-	var processTagValues func(data *map[string]any) error
-	processTagValues = func(data *map[string]any) error {
-		for key, value := range *data {
-			if stringVal, ok := value.(string); ok {
-				// Check for files
-				if isImage, err := isImageFilePath(stringVal); err != nil {
-					return err
-				} else {
-					if isImage {
-						image, err := CreateInlineImage(stringVal)
-						if err != nil {
-							return err
-						}
-						imageXml, err := d.addInlineImage(image)
-						if err != nil {
-							return err
-						}
-						(*data)[key] = imageXml
-					}
-				}
-			} else if nestedMap, ok := value.(map[string]any); ok {
-				if err := processTagValues(&nestedMap); err != nil {
-					return err
-				}
-				(*data)[key] = nestedMap
-			} else if sliceValue, ok := value.([]map[string]any); ok {
-				for i := range sliceValue {
-					if err := processTagValues(&sliceValue[i]); err != nil {
-						return err
-					}
-				}
-			} else if inlineImage, ok := value.(*InlineImage); ok {
-				imageXml, err := d.addInlineImage(inlineImage)
-				if err != nil {
-					return err
-				}
-				(*data)[key] = imageXml
-			}
-		}
-
-		return nil
-	}
-
-	err = processTagValues(&convertedData)
-	if err != nil {
-		return nil, err
-	}
-
-	return convertedData, nil
-}
-
-func isFilePath(filepath string) (bool, error) {
+func IsFilePath(filepath string) (bool, error) {
 	// Check if the path exists
 	if _, err := os.Stat(filepath); err != nil {
 		if os.IsNotExist(err) {
@@ -80,7 +23,7 @@ func isFilePath(filepath string) (bool, error) {
 	return false, nil
 }
 
-func isImageFilePath(filepath string) (bool, error) {
+func IsImageFilePath(filepath string) (bool, error) {
 	ext := path.Ext(filepath)
 	validExts := []string{".png", ".jpg", ".jpeg"}
 	isValid := slices.Contains(validExts, ext)
@@ -88,7 +31,7 @@ func isImageFilePath(filepath string) (bool, error) {
 		return false, nil
 	}
 
-	isFile, err := isFilePath(filepath)
+	isFile, err := IsFilePath(filepath)
 	if err != nil {
 		return false, err
 	}
